requirements.txt
asgiref >= 3.5.2
Django >= 4.1.2
image >= 1.5.33
Pillow >= 9.2.0
pip >= 22.2.2
qrcode >= 7.3.1
setuptools >= 63.4.3
six >= 1.16.0
sqlparse >= 0.4.3
